Added warm thick contribution model
